requirements.txt
pymongo
langdetect
langcodes
language_data
justext
publicsuffixlist
newspaper3k
goose3
whois
BeautifulSoup4
brotli
dateparser
python-dateutil
Unidecode
reppy<=0.4.14
scikit-learn
url_normalize
cchardet
numpy<=1.23.5
pandas<=1.5.3
regex
numexpr<=2.7.3
pytest
sklearn<=0.23.2

azure-identity
azure-keyvault-secrets
azure-storage-queue
azure-storage-file-share

subprocess.run